chore: Move file to service package
