+++ src/main/kotlin/smalltalk/backend/presentation/controller/ChatRoomController.kt
@@ -4,11 +4,11 @@ import org.springframework.messaging.handler.annotation.DestinationVariable
 import org.springframework.messaging.handler.annotation.MessageMapping
 import org.springframework.stereotype.Controller
 import smalltalk.backend.presentation.dto.chatmessage.ChatMessage
-import smalltalk.backend.ChatService
+import smalltalk.backend.application.service.chatroom.ChatRoomService
 
 @Controller
 class ChatRoomController (
-    private val chattingService: ChatService
+    private val chattingService: ChatRoomService
 ) {
 
     @MessageMapping("/{roomId}")
